Member deletion fixed, mostly. TO DO: fix the build to tweak zope.conf, and note the change in CHANGES.txt.
Problem was, the default behavior on deletion in CMF/Plone is to crawl
the entire site to remove local roles, which can be extremely slow
(many minutes on openplans.org).  We now do it asynchronously.  (This
has some risks, but I think it's less risky than the security issues
from leaving the local roles in place.)



git-svn-id: https://svn.socialplanning.org/svn/opencore/trunk@24024 40f73b84-6177-4a8e-8c61-7eee4b1974ed