merge -r 26901:HEAD branches/team-admin-buttons
 * add "Promote admin" and "Demote admin" buttons to manage-team page since apparently nobody notices 
the liveEdit-style dropdowns
 * rewrite set_roles() to be *slightly* less arcane
 * using jQuery instead of oc-behaviors/ext to build ajax ui
  * build and destroy "live edit" DOM elements dynamically (instead of toggling visibility on divs vs 
inputs)
  * after successful change, build new row to swap in entirely on client side, instead of relying on 
server to send back an html snippet and an action
  * still using OC.parseJSON and OC.psm though (nothing wrong with them)

The jQuery stuff is less generic than the oc-behaviors, but it's also less magic, and requires a lot 
less ugly serverside code.  Hoping that I can find a good balance over time as I do this to other 
views.
 



git-svn-id: https://svn.socialplanning.org/svn/opencore/trunk@27068 40f73b84-6177-4a8e-8c61-7eee4b1974ed