Add experimental script for porting a wiki history from CMFEditions to a bzr repository.
A single bzr repository corresponds to a single project wiki.  (NOT an individual wiki page.)

Commits are properly ordered in time based on the original commit date, across the entire repository.

The original timestamp and committer are stored as revprops on the commits,
rather than trying to override the system-level author and timestamp annotations.

Because a project can have a very large number of wiki pages with a very large number of edits per page,
and because CMFEditions uses a CVS-like model of local revision numbers instead of an SVN-like global revision number,
sorting all the wiki pages by time can be an extremely memory-intensive operation.  We approach this by storing
the individual checkins' essential metadata in a temporary sqlite db, and deferring the sorting to the database.
So it's a two-phase process: first, import the checkins from CMFEditions to a sqlite db; then, import them from the db
to a bzr repository.

Not sure how well this will work with real data, going to try that next.



git-svn-id: https://svn.socialplanning.org/svn/opencore/trunk@27849 40f73b84-6177-4a8e-8c61-7eee4b1974ed